# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrustDesk API - Verifiable support platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface User {
  id: string;
  primaryWallet?: string | null;
  primaryEmail?: string | null;
  reputationScore: number;
  riskScore: number;
  isStaff: boolean;
  isBanned: boolean;
  createdAt: string;
}

export interface SessionResponse {
  authenticated: boolean;
  user?: User;
}

export interface AuthResponse {
  ok: boolean;
  user?: User;
}

export interface RequestEmailOtpBody {
  email: string;
}

export interface VerifyEmailOtpBody {
  email: string;
  otp: string;
}

export interface NonceResponse {
  nonce: string;
}

export interface VerifyWalletBody {
  message: string;
  signature: string;
  address: string;
}

export interface RiskEvent {
  id: string;
  eventType: string;
  scoreDelta?: number | null;
  createdAt: string;
}

export type UserWithRisk = User & {
  recentRiskEvents: RiskEvent[];
};

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  OPEN: "OPEN",
  IN_PROGRESS: "IN_PROGRESS",
  RESOLVED: "RESOLVED",
  CLOSED: "CLOSED",
} as const;

export type TicketPriority =
  (typeof TicketPriority)[keyof typeof TicketPriority];

export const TicketPriority = {
  LOW: "LOW",
  NORMAL: "NORMAL",
  HIGH: "HIGH",
  URGENT: "URGENT",
} as const;

export interface Ticket {
  id: string;
  creatorId: string;
  status: TicketStatus;
  subject?: string | null;
  priority: TicketPriority;
  riskScore: number;
  createdAt: string;
  updatedAt: string;
  lastMessageAt?: string | null;
  creator?: User;
}

export interface TicketListResponse {
  tickets: Ticket[];
  total: number;
  page: number;
  limit: number;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  USER: "USER",
  STAFF: "STAFF",
  SYSTEM: "SYSTEM",
} as const;

export interface MessageSignature {
  id: string;
  signature: string;
  publicKey: string;
  signatureType: string;
  createdAt: string;
}

export interface Message {
  id: string;
  ticketId: string;
  senderId: string;
  role: MessageRole;
  content: string;
  sequenceNumber: number;
  prevHash?: string | null;
  messageHash: string;
  createdAt: string;
  signature?: MessageSignature;
  sender?: User;
}

export type TicketWithMessages = Ticket & {
  messages: Message[];
};

export type CreateTicketBodyPriority =
  (typeof CreateTicketBodyPriority)[keyof typeof CreateTicketBodyPriority];

export const CreateTicketBodyPriority = {
  LOW: "LOW",
  NORMAL: "NORMAL",
  HIGH: "HIGH",
  URGENT: "URGENT",
} as const;

export interface CreateTicketBody {
  subject: string;
  content: string;
  priority?: CreateTicketBodyPriority;
}

export type UpdateTicketBodyStatus =
  (typeof UpdateTicketBodyStatus)[keyof typeof UpdateTicketBodyStatus];

export const UpdateTicketBodyStatus = {
  OPEN: "OPEN",
  IN_PROGRESS: "IN_PROGRESS",
  RESOLVED: "RESOLVED",
  CLOSED: "CLOSED",
} as const;

export type UpdateTicketBodyPriority =
  (typeof UpdateTicketBodyPriority)[keyof typeof UpdateTicketBodyPriority];

export const UpdateTicketBodyPriority = {
  LOW: "LOW",
  NORMAL: "NORMAL",
  HIGH: "HIGH",
  URGENT: "URGENT",
} as const;

export interface UpdateTicketBody {
  status?: UpdateTicketBodyStatus;
  priority?: UpdateTicketBodyPriority;
}

export interface SendMessageBody {
  content: string;
}

export interface IntegrityResult {
  valid: boolean;
  messageCount: number;
  brokenAt?: number | null;
  details?: string;
}

export type DashboardSummaryTicketsByStatus = { [key: string]: number };

export type DashboardSummaryTicketsByPriority = { [key: string]: number };

export interface DashboardSummary {
  openTickets: number;
  inProgressTickets: number;
  resolvedToday: number;
  totalUsers: number;
  highRiskUsers: number;
  avgResponseTimeMinutes: number;
  ticketsByStatus: DashboardSummaryTicketsByStatus;
  ticketsByPriority: DashboardSummaryTicketsByPriority;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  TICKET_CREATED: "TICKET_CREATED",
  MESSAGE_SENT: "MESSAGE_SENT",
  STATUS_CHANGED: "STATUS_CHANGED",
  USER_BANNED: "USER_BANNED",
  TICKET_RESOLVED: "TICKET_RESOLVED",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  ticketId?: string | null;
  userId?: string | null;
  description: string;
  timestamp: string;
}

export interface ActivityFeed {
  items: ActivityItem[];
}

export interface RiskOverview {
  highRiskCount: number;
  mediumRiskCount: number;
  lowRiskCount: number;
  recentEvents: RiskEvent[];
  flaggedTickets: number;
}

export type StaffActionMetadata = { [key: string]: unknown } | null;

export interface StaffAction {
  id: string;
  staffId?: string | null;
  ticketId?: string | null;
  actionType: string;
  metadata?: StaffActionMetadata;
  createdAt: string;
  staff?: User;
}

export interface StaffActionList {
  actions: StaffAction[];
}

export type ListTicketsParams = {
  status?: ListTicketsStatus;
  page?: number;
  limit?: number;
};

export type ListTicketsStatus =
  (typeof ListTicketsStatus)[keyof typeof ListTicketsStatus];

export const ListTicketsStatus = {
  OPEN: "OPEN",
  IN_PROGRESS: "IN_PROGRESS",
  RESOLVED: "RESOLVED",
  CLOSED: "CLOSED",
} as const;

export type GetRecentActivityParams = {
  limit?: number;
};

export type ListStaffActionsParams = {
  limit?: number;
};