# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrustDesk API - Verifiable support platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/email/request-otp:
    post:
      operationId: requestEmailOtp
      tags: [auth]
      summary: Request email OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestEmailOtpBody"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /auth/email/verify-otp:
    post:
      operationId: verifyEmailOtp
      tags: [auth]
      summary: Verify email OTP and authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailOtpBody"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/wallet/verify:
    post:
      operationId: verifyWalletSignature
      tags: [auth]
      summary: Verify SIWE signature and authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyWalletBody"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List tickets (user sees own, staff sees all)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketListResponse"

    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a new support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketBody"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{ticketId}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get a ticket by ID
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketWithMessages"

    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket (staff only)
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketBody"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{ticketId}/messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a ticket
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /tickets/{ticketId}/verify:
    get:
      operationId: verifyTicketIntegrity
      tags: [tickets]
      summary: Verify the hash chain integrity of a ticket
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Integrity verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrityResult"

  # Users
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    AuthResponse:
      type: object
      properties:
        ok:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
      required: [ok]
    RequestEmailOtpBody:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    VerifyEmailOtpBody:
      type: object
      properties:
        email:
          type: string
          format: email
        otp:
          type: string
      required: [email, otp]
...
    VerifyWalletBody:
      type: object
      properties:
        message:
          type: string
        signature:
          type: string
        address:
          type: string
      required: [message, signature, address]
    User:
      type: object
      properties:
        id:
          type: string
        primaryWallet:
          type: string
          nullable: true
        primaryEmail:
          type: string
          nullable: true
        reputationScore:
          type: number
        riskScore:
          type: number
        isStaff:
          type: boolean
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, reputationScore, riskScore, isStaff, isBanned, createdAt]
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: string
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        subject:
          type: string
          nullable: true
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
        riskScore:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
        creator:
          $ref: "#/components/schemas/User"
      required: [id, creatorId, status, priority, riskScore, createdAt, updatedAt]
    TicketListResponse:
      type: object
      properties:
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [tickets, total, page, limit]
    TicketWithMessages:
      allOf:
        - $ref: "#/components/schemas/Ticket"
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"
          required: [messages]
    CreateTicketBody:
      type: object
      properties:
        subject:
          type: string
        content:
          type: string
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
      required: [subject, content]
    UpdateTicketBody:
      type: object
      properties:
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
    Message:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        senderId:
          type: string
        role:
          type: string
          enum: [USER, STAFF, SYSTEM]
        content:
          type: string
        sequenceNumber:
          type: integer
        prevHash:
          type: string
          nullable: true
        messageHash:
          type: string
        createdAt:
          type: string
          format: date-time
        signature:
          $ref: "#/components/schemas/MessageSignature"
        sender:
          $ref: "#/components/schemas/User"
      required: [id, ticketId, senderId, role, content, sequenceNumber, messageHash, createdAt]
    MessageSignature:
      type: object
      properties:
        id:
          type: string
        signature:
          type: string
        publicKey:
          type: string
        signatureType:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, signature, publicKey, signatureType, createdAt]
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
    IntegrityResult:
      type: object
      properties:
        valid:
          type: boolean
        messageCount:
          type: integer
        brokenAt:
          type: integer
          nullable: true
        details:
          type: string
      required: [valid, messageCount]
...