# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrustDesk API - Verifiable support platform
 * OpenAPI spec version: 0.1.0
 */
import type { TicketPriority } from "./ticketPriority";
import type { TicketStatus } from "./ticketStatus";
import type { User } from "./user";

export interface Ticket {
  id: string;
  creatorId: string;
  status: TicketStatus;
  subject?: string | null;
  priority: TicketPriority;
  riskScore: number;
  createdAt: Date;
  updatedAt: Date;
  lastMessageAt?: Date | null;
  creator?: User;
}