# lib/api-zod/src/generated/types/updateTicketBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrustDesk API - Verifiable support platform
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTicketBodyPriority } from "./updateTicketBodyPriority";
import type { UpdateTicketBodyStatus } from "./updateTicketBodyStatus";

export interface UpdateTicketBody {
  status?: UpdateTicketBodyStatus;
  priority?: UpdateTicketBodyPriority;
}