# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrustDesk API - Verifiable support platform
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  primaryWallet?: string | null;
  primaryEmail?: string | null;
  reputationScore: number;
  riskScore: number;
  isStaff: boolean;
  isBanned: boolean;
  createdAt: Date;
}